Finished UI and add some requred to the field

+++ CookingBook/Pages/EditPage.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using CookingBook.Services;
 using CookingBook.Models;
+using System.Diagnostics;
 
 namespace CookingBook.Pages;
 
@@ -42,46 +43,40 @@ public partial class EditPage : ContentPage
     {
         if (_recipe != null)
         {
-            await App.RecipeRepo.DeleteRecipe(_recipe.id);
-            await DisplayAlert("Success", "Recipe deleted successfully.", "OK");
+            bool answer = await DisplayAlert("Question?", "Are you sure you want to DELETE recipe?", "Yes", "No");
+            Debug.WriteLine("Answer: "+ answer);
+            if (answer is true)
+            {
+                await App.RecipeRepo.DeleteRecipe(_recipe.id);
+                await Shell.Current.GoToAsync(nameof(ListPage));
+            }
 
-            await Shell.Current.GoToAsync(nameof(ListPage));
+            
 
         }
     }
 
     private async void btnUpDate_Clicked(object sender, EventArgs e)
     {
-        try
-        {
+       
             // Оновлюємо рецепт за допомогою методу з репозиторію
             _recipe = await App.RecipeRepo.UpdateRecipe(recId, _recipe);
 
-            if (_recipe != null)
-            {
-                // Оновлюємо дані рецепта зі значеннями з полів введення
-                _recipe.Name = entryName.Text;
-                _recipe.Description = entryDescr.Text;
-
-                // Фактично оновлюємо рецепт у базі даних після оновлення властивостей
-                await App.RecipeRepo.UpdateRecipe(recId, _recipe);
-
-                // Повідомляємо про успішне оновлення
-                await DisplayAlert("Success", "Recipe updated successfully.", "OK");
-
-                // Переходимо на сторінку зі списком після успішного оновлення
-                await Shell.Current.GoToAsync(nameof(ListPage));
-            }
-            else
-            {
-                // Можна обробити ситуацію, коли оновлення не було успішним
-                await DisplayAlert("Error", "Failed to update recipe.", "OK");
-            }
-        }
-        catch (Exception ex)
+        if (_recipe != null)
         {
-            // Обробляємо будь-які винятки, що можуть виникнути під час оновлення
-            await DisplayAlert("Error", $"An error occurred: {ex.Message}", "OK");
+            // Оновлюємо дані рецепта зі значеннями з полів введення
+            _recipe.Name = entryName.Text;
+            _recipe.Description = entryDescr.Text;
+
+            // Фактично оновлюємо рецепт у базі даних після оновлення властивостей
+            await App.RecipeRepo.UpdateRecipe(recId, _recipe);
+
+            // Повідомляємо про успішне оновлення
+            await DisplayAlert("Success", "Recipe updated successfully.", "OK");
+
+            // Переходимо на сторінку зі списком після успішного оновлення
+            await Shell.Current.GoToAsync(nameof(ListPage));
+
         }
     }
 }
